+++ lib/src/budgets/budget_controller.dart
@@ -2,11 +2,15 @@ import 'package:uuid/uuid.dart';
 
 import 'package:flutter_xspend/src/models/budget.dart';
 import 'package:flutter_xspend/src/models/user.dart';
-import 'package:flutter_xspend/src/helpers/budget_helper.dart';
 
 class BudgetController {
   static bool isValidForm(name, amount, startDate, endDate) {
     return name != null && name != '' && amount != null && amount != '' && double.parse(amount) > 0 && startDate != null && endDate != null;
+  }
+
+  static loadBudgets(callback) async {
+    final List<Budget> budgets = await Budget.getAllOfCurrentUser();
+    callback?.call(budgets);
   }
 
   static create(name, amount, startDate, endDate, currencyType, callback) async {
@@ -21,7 +25,7 @@ class BudgetController {
                     ..user.value = await User.currentLoggedIn();
 
     Budget.create(budget);
-    BudgetHelper.loadBudgets((budgets) {
+    loadBudgets((budgets) {
       callback?.call(budgets);
     });
   }
@@ -36,7 +40,7 @@ class BudgetController {
     };
     Budget.update(id, params);
     Future.delayed(const Duration(milliseconds: 100), () {
-      BudgetHelper.loadBudgets((budgets) {
+      loadBudgets((budgets) {
         callback?.call(budgets);
       });
     });
@@ -45,7 +49,7 @@ class BudgetController {
   static delete(String id, callback) {
     Budget.deleteById(id);
     Future.delayed(const Duration(milliseconds: 100), () {
-      BudgetHelper.loadBudgets((budgets) {
+      loadBudgets((budgets) {
         callback?.call(budgets);
       });
     });
